drivers projects: fix CodeQL-flagged defects
Resolve static-analysis findings reported by CodeQL across drivers and
example projects. All fixes are behavior-preserving except where the
previous behavior was itself a bug (sd.c, adin1110 ethertype print).

- sd-card: sd.c: guard against uint32_t underflow when response_len is
  0 (compare response_len > 1 instead of response_len - 1 > 0), which
  previously passed a huge length to memset.

- Widen narrow loop indices compared against wider bounds so the loop
  cannot wrap: ad5706r, ad9517, ad9523, ad9528, max96792_csi,
  iio_adis, ade9113, adpd410x, stm32 dma/gpdma/spi/xspi, nmea_ubx,
  eval-adxl367z dummy_example.

- Match printf/snprintf format specifiers to their argument types
  (cast uint32_t/int32_t args to (unsigned) long for %l): ad9144,
  ad9523, nmea_ubx, iio_adis, ad413x, ad469x_evb, ad713x_fmcz,
  ad796x_fmcz, ad9656_fmc, aducm3029_flash_demo, cn0561, cn0565,
  pulsar-adc. ad413x dropped a stray argument; adin1110 printed a
  two-byte ethertype array with a single %02X - now prints both bytes.

- Cast an operand to the wider type before multiplying to avoid
  overflow before the widening conversion: iio_dac_demo, ad9545,
  adf4156, adf4157, adv7511 cf_hdmi.

- cn0565 eit: check sscanf return against the expected conversion
  count (== 1); the uint8_t result stored EOF as 255, so parse
  failures were treated as success.

+++ drivers/dac/ad9144/ad9144.c
@@ -1304,8 +1304,8 @@ int32_t ad9144_short_pattern_test(struct ad9144_dev *dev,
 						      0x01, 0x00);
 			if (ret == -1)
 				printf("%s : short-pattern-test mismatch (%#06lx, %#06lx, %#06lx)!.\n",
-				       __func__, dac, sample,
-				       init_param->stpl_samples[dac][sample]);
+				       __func__, (unsigned long)dac, (unsigned long)sample,
+				       (unsigned long)init_param->stpl_samples[dac][sample]);
 		}
 	}
 	return 0;
